rag/requirements.txt
python-dotenv
langchain
fastapi[all]
octoai-sdk
openai
langchain-pinecone
langchain-openai
pinecone-client
pypdf
datasets